A LOT  of changes.
Trying to fix small errors as they pop up. Biggest one right now is that
the arm jitters in.

+++ old/LiftFunctions.h
@@ -32,7 +32,7 @@ void armBump(int direction)
 }
 void boostControl(int type, bool speedToggle, int drive, int speedMod)
 {
-	const int max = 127;
+	const byte max = 127;
 	if(speedToggle)
 	{
 		drive = (drive*1.2);
